% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-programming.R
\name{audit_wrap}
\alias{audit_wrap}
\alias{audit_peek}
\alias{audit_poke}
\alias{audit_unwrap}
\title{A lightweight container for auditing or logging}
\usage{
audit_wrap(data)

audit_peek(x, fn, ...)

audit_poke(x, fn, ...)

audit_unwrap(x)
}
\arguments{
\item{data}{data to wrap into an audit}

\item{x}{an audit object}

\item{fn}{a function to apply to the data inside of the audit object}

\item{...}{further arguments for \code{fn}}
}
\value{
\code{audit_wrap()}, \code{audit_peek()}, \code{audit_poke()} return audit objects.
\code{audit_unwrap()} returns the \verb{$data} from an \code{audit}.
}
\description{
An \code{audit} is a \code{list()} with two fields:
\enumerate{
\item \code{data}: most recently updated (or poked) data
\item \code{notes}: a history of the \code{fn} calls and results applied to the object.
}

Usage:
\itemize{
\item \code{audit_wrap(data)} wraps some data inside of an audit.
\item \code{audit_peek(x, fn, ...)} applies a function to the data, log the results
in the \code{notes}, but \emph{not} overwrite the audit data.
\item \code{audit_poke(x, fn, ...)} applies a function to the data, log the results
in the \code{notes}, and overwrite the audit data.
\item \code{audit_unwrap()} returns the underlying data.
}
}
\details{
There is no \code{audit_notes()} function at this time because it is not
clear what the most convenient or tidy format would be for the notes.

These container was designed for piping a vector of filenames through a
series of data validation functions. Therefore, the intended use case is a
vector of a data and not larger objects.
}
\examples{
x <- letters |>
  audit_wrap() |>
  audit_peek(rev) |>
  audit_peek(toupper)

# Note that none of the underlying data has changed. This is useful for
# finding values that fail validation checks.
x

# Use _poke() to update the data:
x <- letters |>
  audit_wrap() |>
  audit_peek(rev) |>
  audit_poke(toupper) |>
  audit_peek(getElement, 1) |>
  audit_peek(function(x) character(0))
x

# Index into the notes to pull the data
x$notes[[3]][["result"]]
}
\concept{programming-utils}
