% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_overlap_rate}
\alias{compute_overlap_rate}
\title{Compute overlap rate for (phoneme alignment) intervals}
\usage{
compute_overlap_rate(x1, x2, y1, y2)
}
\arguments{
\item{x1, x2}{start and end times for the first interval}

\item{y1, y2}{start and end times for the second interval}
}
\value{
the overlap rate
}
\description{
Compute overlap rate for (phoneme alignment) intervals
}
\details{
Paulo and Oliveira (2004) provide an "overlap rate" statistic for computing
the amount of overlap between two (time) intervals. To my knowledge, nobody
has described the Overlap Rate in this way, but it is the
\href{https://en.wikipedia.org/wiki/Jaccard_index}{Jaccard index} applied to time
intervals.

Let \eqn{X=[x_\text{min}, x_\text{max}]} and
\eqn{Y=[y_\text{min}, y_\text{max}]} be the sets of times spanned by the
intervals \eqn{x} and \eqn{y}. Then, \eqn{X \cap Y} is the \emph{intersection} or the
times covered by both intervals, and \eqn{X \cup Y} is the \emph{union} or the
times covered by either interval. The size of a set \eqn{A} is denoted
\eqn{|A|}. Then the overlap rate is the Jaccard index or the proportion of
elements that the two sets have in common:

\deqn{\text{overlap rate} = \frac{|X \cap Y|}{|X \cup Y|}}
}
\examples{
compute_overlap_rate(
  c(0.0, 0.0, 0.0, 0.0),
  c(1.0, 1.0, 1.0,  NA),
  c(0.5, 2.0, 1.0, 1.0),
  c(2.0, 3.0, 2.0, 2.0)
)
}
\references{
Paulo, S., & Oliveira, L. C. (2004). Automatic Phonetic
Alignment and Its Confidence Measures. In J. L. Vicedo, P. Martínez-Barco,
R. Muńoz, & M. Saiz Noeda (Eds.), \emph{Advances in Natural Language Processing}
(pp. 36–44). Springer. \url{https://doi.org/10.1007/978-3-540-30228-5_4}
}
