% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_value.R
\name{compute_predictive_value_from_rates}
\alias{compute_predictive_value_from_rates}
\title{Compute positive and negative predictive value}
\usage{
compute_predictive_value_from_rates(sensitivity, specificity, prevalence)
}
\arguments{
\item{sensitivity, specificity, prevalence}{vectors of confusion matrix rates}
}
\value{
a tibble with the columns \code{sensitivity}, \code{specificity}, \code{prevalence},
\code{ppv}, \code{npv} where \code{ppv} and \code{npv} are the positive predictive value and
the negative predictive value.
}
\description{
Compute positive and negative predictive value
}
\details{
These vectors passed into this function should be some common length
and/or length 1. For example, 4 sensitivities, 4 specificities and 1
incidence will work because the sensitivities and specificities have a common
length and we can safely recycle (reuse) the incidence value. But 4
sensitivities, 2 specificities, and 1 incidence will \emph{fail} because there is
not a common length.
}
\examples{
compute_predictive_value_from_rates(
  sensitivity = .9,
  specificity = .8,
  prevalence = .05
)

compute_predictive_value_from_rates(
  sensitivity = .67,
  specificity = .53,
  prevalence = c(.15, .3)
)
}
\concept{roc}
