% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fake_intelligibility}
\alias{data_fake_intelligibility}
\title{Fake intelligibility data}
\format{
A data frame with 200 rows and 2 variables:
\describe{
\item{age_months}{child's age in months}
\item{intelligibility}{child's intelligibility (proportion of words said
by the child that were correctly transcribed by two listeners)}
}
}
\usage{
data_fake_intelligibility
}
\description{
A dataset of fake intelligibility scores for testing and demonstrating
modeling functions. These were created by randomly sampling 200 rows of an
intelligibility dataset and adding random noise to the \code{age_months} and
\code{intelligibility} variables. These values do not measure any real children
but represent plausible age and intelligibility measurements from our kind of
work.
}
\concept{datasets}
\keyword{datasets}
