% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fake_rates}
\alias{data_fake_rates}
\title{Fake speaking rate data}
\format{
A data frame with 200 rows and 2 variables:
\describe{
\item{age_months}{child's age in months}
\item{speaking_sps}{child's speaking rate in syllables per second}
}
}
\usage{
data_fake_rates
}
\description{
A dataset of fake speaking rate measures for testing and demonstrating
modeling functions. These were created by randomly
sampling 200 rows of a speaking rate dataset and adding random noise to the
\code{age_months} and \code{speaking_sps} variables. These values do not measure any
real children but represent plausible age and rate measurements from our
kind of work.
}
\concept{datasets}
\keyword{datasets}
