% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{file_replace_name}
\alias{file_replace_name}
\alias{file_rename_with}
\title{Rename file basenames using functions}
\usage{
file_replace_name(
  path,
  pattern,
  replacement,
  .dry_run = TRUE,
  .overwrite = FALSE
)

file_rename_with(path, .fn, ..., .dry_run = TRUE, .overwrite = FALSE)
}
\arguments{
\item{path}{vector of paths for files to rename}

\item{pattern, replacement}{arguments forwarded to \code{\link[stringr:str_replace]{stringr::str_replace()}}}

\item{.dry_run}{when \code{FALSE}, files are renamed. When \code{TRUE} (the default),
no files are renamed but the rename plan is printed.}

\item{.overwrite}{Whether to overwrite files. Defaults to \code{FALSE} so that
overwriting files is opt-in.}

\item{.fn}{function to apply to file basenames}

\item{...}{arguments passed onto \code{.fn}}
}
\value{
a dataframe describing the file-renaming plan. Files that would
overwrite an existing file are skipped unless \code{.overwrite = TRUE}. This
function throws an error if a name collision is detected, where two files
are both renamed into the same target path.
}
\description{
\code{file_replace_name()} uses \code{\link[stringr:str_replace]{stringr::str_replace()}} to rename files.
\code{file_rename_with()} allows you to rename files with a generic
string-transforming function.
}
\details{
Only the basename of the file (returned by \code{\link[=basename]{basename()}})
undergoes string replacement.
}
\examples{
# With .dry_run = TRUE, we can make up some file paths.
dir <- "//some-fake-location/"
path <- file.path(
  dir,
  c("report_1.csv", "report_2.csv", "report-1.csv", "skipped.csv")
)

updated <- file_replace_name(path, "report_", "report-", .dry_run = TRUE)

# Collisions are detected
updated <- file_replace_name(path, "report_\\\\d", "report-1", .dry_run = TRUE)

# Doing nothing
updated <- file_rename_with(path, identity, .dry_run = TRUE)
}
