% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rate.R
\name{fit_gen_gamma_gamlss}
\alias{fit_gen_gamma_gamlss}
\alias{predict_gen_gamma_gamlss}
\title{Fit a generalized gamma regression model (for speaking rate)}
\usage{
fit_gen_gamma_gamlss(data, var_x, var_y, df_mu = 3, df_sigma = 2, df_nu = 1)

predict_gen_gamma_gamlss(newdata, model, centiles = c(5, 10, 50, 90, 95))
}
\arguments{
\item{data}{a data frame}

\item{var_x, var_y}{(unquoted) variable names giving of the predictor variable
(e.g., \code{age}) and outcome variable (.e.g, \code{rate}).}

\item{df_mu, df_sigma, df_nu}{degrees of freedom}

\item{newdata}{a one-column dataframe for predictions}

\item{model}{a model fitted by \code{\link[=fit_gen_gamma_gamlss]{fit_gen_gamma_gamlss()}}}

\item{centiles}{centiles to use for prediction. Defaults to \code{c(5, 10, 50, 90, 95)}.}
}
\value{
for \code{fit_gen_gamma_gamlss()}, a \code{\link[=mem_gamlss]{mem_gamlss()}}-fitted model. The
\code{.user} data in the model includes degrees of freedom for each parameter
and the \code{\link[splines:ns]{splines::ns()}} basis for each parameter. For
\code{predict_gen_gamma_gamlss()}, a dataframe containing the model predictions
for mu, sigma, and nu, plus columns for each centile in \code{centiles}.
}
\description{
The function fits the same type of model as used in Mahr 2021: A generalized
gamma regression model with natural cubic splines on the mean (mu), scale
(sigma), and shape (nu). This model is fitted using \code{\link[=mem_gamlss]{mem_gamlss()}}.
}
\details{
\code{\link[=predict_centiles]{predict_centiles()}} will work with this function, but it will likely throw a
warning message. Therefore, \code{\link[=predict_gen_gamma_gamlss]{predict_gen_gamma_gamlss()}} provides an
alternative way to compute centiles from the model. It manually computes the
centiles instead of relying on \code{\link[gamlss:centiles]{gamlss::centiles()}}.
}
\examples{
data_fake_rates

m <- fit_gen_gamma_gamlss(data_fake_rates, age_months, speaking_sps)
summary(m)

# Create growth curves for a sub age-range
new_ages <- data.frame(age_months = 48:71)
centiles <- predict_gen_gamma_gamlss(new_ages, m)
centiles

if(requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  ggplot(pivot_centiles_longer(centiles)) +
    aes(x = age_months, y = .value) +
    geom_line(aes(group = .centile, color = .centile_pair)) +
    geom_point(
      aes(y = speaking_sps),
      data = subset(
        data_fake_rates,
        48 <= age_months & age_months <= 71
      )
    )
}

}
\concept{models}
