% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{logitnorm_mean}
\alias{logitnorm_mean}
\title{Compute the mean of logit-normal distribution(s)}
\usage{
logitnorm_mean(mu, sigma)
}
\arguments{
\item{mu}{mean(s) on the logit scale}

\item{sigma}{standard deviation(s) on the logit scale}
}
\value{
the means of the distributions
}
\description{
This function is a wrapper around \code{\link[logitnorm:momentsLogitnorm]{logitnorm::momentsLogitnorm()}}.
}
\examples{
\donttest{
x <- logitnorm_mean(2, 1)
x
}
# compare to simulation
set.seed(100)
rnorm(1000, 2, 1) |> plogis() |> mean()
}
