% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss-centiles.R
\name{mem_gamlss}
\alias{mem_gamlss}
\title{Fit a gamlss model but store user data}
\usage{
mem_gamlss(...)
}
\arguments{
\item{...}{arguments passed to gamlss::gamlss()}
}
\value{
the fitted `model` object but updated to include user information in
  `model$.user`. Includes the dataset used to fit the model
  `model$.user$data`, the session info `model$.user$session_info` and the
  call used to fit the model `model$.user$call`. `model$call` is updated to
  match
}
\description{
Think of it as a gamlss model with memories (mem. gamlss).
}
