% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss-centiles.R
\name{predict_centiles}
\alias{predict_centiles}
\alias{pivot_centiles_longer}
\title{Predict and tidy centiles from a GAMLSS model}
\usage{
predict_centiles(newdata, model, centiles = c(5, 10, 50, 90, 95), ...)

pivot_centiles_longer(data)
}
\arguments{
\item{newdata}{a one-column dataframe for predictions}

\item{model}{a gamlss model prepared by \code{mem_gamlss()}}

\item{centiles}{centiles to use for prediction. Defaults to \code{c(5, 10, 50, 90, 95)}.}

\item{...}{arguments passed to \code{gamlss::centiles.pred()}}

\item{data}{centile predictions to reshape for \code{pivot_centiles_longer()}}
}
\value{
a tibble with fitted centiles for \code{predict_centiles()} and a
long-format tibble with one centile value per row in
\code{pivot_centiles_longer()}
}
\description{
gamlss has trouble doing predictions without the original training data.
}
\concept{gamlss}
