% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{print_duckdb}
\alias{print_duckdb}
\title{Custom function for printing duckdb database connections}
\usage{
print_duckdb(object, table_limit = 50)
}
\arguments{
\item{object}{a database connection to a duckdb database}

\item{table_limit}{number of table names to print before truncating with
\code{...}. Defaults to 50.}
}
\value{
\code{NULL} invisibly
}
\description{
Custom function for printing duckdb database connections
}
\details{
Use the following to overwrite the S4 method for printing duckdb
objects.

\if{html}{\out{<div class="sourceCode">}}\preformatted{setMethod("show", "duckdb_connection", wisclabmisc::print_duckdb)
}\if{html}{\out{</div>}}
}
\concept{database}
