% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{tbl_bind}
\alias{tbl_bind}
\title{Select and row-bind multiple database tables from the same source together}
\usage{
tbl_bind(src, table_selection, id_name = ".source")
}
\arguments{
\item{src}{a database connection}

\item{table_selection}{a \code{\link[tidyselect:language]{tidy-select}} name
selection. For example, \code{starts_with("loc_")} would select all tables that
start with \code{"loc_"}. Place multiple selections inside of \code{c()}.}

\item{id_name}{Name for the table ID column. The names of the original tables
will be included in a column named by \code{id_name}. By default, this column is
\code{".source"}. If \code{NULL}, then the ID column is not included.}
}
\value{
a tbl (query to a remote table)
}
\description{
Select and row-bind multiple database tables from the same source together
}
\details{
In dplyr, \code{tbl(src, table_name)} names and queries a database table.
\verb{tbl_bind(src, <selection>)} extends this idea into selecting and
combining multiple database tables.

To manually combine two tbls \code{x} and \code{y}, use \code{\link[dplyr:union_all]{dplyr::union_all(x, y)}}.
}
\examples{
library(dplyr)
db <- duckdb::duckdb() |>
  DBI::dbConnect() |>
  withr::local_db_connection()

DBI::dbWriteTable(db, "mtcars_g1", mtcars[mtcars$cyl == 4, ])
DBI::dbWriteTable(db, "mtcars_g2", mtcars[mtcars$cyl == 6, ])
DBI::dbWriteTable(db, "mtcars_g3", mtcars[mtcars$cyl == 8, ])
DBI::dbWriteTable(db, "trees", trees)

DBI::dbListTables(db)

r <- db |>
  tbl_bind(starts_with("mtcars")) |>
  count(.source)
r

show_query(r)
}
\concept{database}
