% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tocs_item}
\alias{tocs_item}
\alias{tocs_type}
\alias{tocs_length}
\title{Extract the TOCS details from a string (usually a filename)}
\usage{
tocs_item(xs)

tocs_type(xs)

tocs_length(xs)
}
\arguments{
\item{xs}{a character vector}
}
\value{
\code{tocs_item()} returns the substring with the TOCS item, \code{tocs_type()}
returns whether the item is \code{"single-word"} or \code{"multiword"}, and
\code{tocs_length()} returns the length of the TOCS item (i.e., the number of
words).
}
\description{
Extract the TOCS details from a string (usually a filename)
}
\examples{
x <- c(
  "XXv16s7T06.lab", "XXv15s5T06.TextGrid", "XXv13s3T10.WAV",
  "XXv18wT11.wav", "non-matching", "s2T01",
  "XXv01s4B01.wav", "XXv01wB01.wav",
  # sometimes these have tags for *v*irtual visits or recording attempts
  "XXv13s3T10v.WAV", "XXv13s3T10a.lab"
)
data.frame(
  x = x,
  item = tocs_item(x),
  type = tocs_type(x),
  length = tocs_length(x)
)
}
\concept{data-utils}
