% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{trapezoid_auc}
\alias{trapezoid_auc}
\alias{partial_trapezoid_auc}
\title{Compute AUCs using the trapezoid method}
\usage{
trapezoid_auc(xs, ys)

partial_trapezoid_auc(xs, ys, xlim)
}
\arguments{
\item{xs, ys}{x and y positions}

\item{xlim}{two-element vector (a range) of the \code{xs} to sum
over}
}
\value{
the area under the curve computed using the trapezoid method. For
\code{partial_trapezoid_auc()}, the partial area under the curve is computed.
}
\description{
Compute AUCs using the trapezoid method
}
\examples{
# Predict whether a car has automatic vs. manual transmission from mpg
r <- pROC::roc(am ~ mpg, mtcars)
pROC::auc(r)
trapezoid_auc(r$specificities, r$sensitivities)

pROC::auc(r, partial.auc = c(.9, 1), partial.auc.focus = "sp")
partial_trapezoid_auc(r$specificities, r$sensitivities, c(.9, 1))

pROC::auc(r, partial.auc = c(.9, 1), partial.auc.focus = "se")
partial_trapezoid_auc(r$sensitivities, r$specificities, c(.9, 1))

pROC::auc(r, partial.auc = c(.1, .9), partial.auc.focus = "sp")
partial_trapezoid_auc(r$specificities, r$sensitivities, c(.1, .9))

pROC::auc(r, partial.auc = c(.1, .9), partial.auc.focus = "se")
partial_trapezoid_auc(r$sensitivities, r$specificities, c(.1, .9))
}
\concept{roc}
