% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisclabmisc-package.R
\docType{package}
\name{wisclabmisc-package}
\alias{wisclabmisc}
\alias{wisclabmisc-package}
\title{wisclabmisc: Tools to Support the 'WiscLab'}
\description{
A collection of 'R' functions for use (and re-use) across 'WiscLab' projects. These are analysis or presentation oriented functions--that is, they are not for data reading or data cleaning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/wisclabmisc}
  \item \url{https://www.tjmahr.com/wisclabmisc/}
  \item Report bugs at \url{https://github.com/tjmahr/wisclabmisc/issues}
}

}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
