% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_year_month_age}
\alias{format_year_month_age}
\title{Convert age in months to years;months}
\usage{
format_year_month_age(x)
}
\arguments{
\item{x}{a vector ages in months}
}
\value{
ages in the years;months format
}
\description{
Convert age in months to years;months
}
\details{
Ages of \code{NA} return \code{"NA;NA"}.

This format by default is not numerically ordered. This means that \code{c("2;0", "10;10", "10;9")} would sort as \code{c("10;10", "10;9", "2;0")}. The function
\code{stringr::str_sort(..., numeric = TRUE)} will sort this vector correctly.
}
\examples{
ages <- c(26, 58, 25, 67, 21, 59, 36, 43, 27, 49)
format_year_month_age(ages)
}
\concept{etc}
