% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_year_month_age}
\alias{format_year_month_age}
\alias{parse_year_month_age}
\alias{parse_yymm_age}
\title{Convert between age in months, years;months, and yymm age formats}
\usage{
format_year_month_age(x, sep = ";")

parse_year_month_age(x, sep = ";")

parse_yymm_age(x, start = 1L)
}
\arguments{
\item{x}{\itemize{
\item \code{format_year_month_age()}: a numeric vector of (non-negative) ages in
months.
\item \code{parse_year_month_age()}: a character vector of ages in \code{"years;months"}
format (or \verb{years\{sep\}months} format more generally).
\item \code{parse_yymm_age()}: a character vector of ages in \code{"yymm"} format.
}}

\item{sep}{Separator to use for \code{year_month} functions. Defaults to \verb{;}.}

\item{start}{For \code{parse_yymm_age()}, the location of the starting
character the \code{yymm} sequence. Defaults to 1.}
}
\value{
\itemize{
\item \code{format_year_month_age()} returns a character vector in \code{"years;months"}
format (or \verb{years\{sep\}months} format more generally).
\item \code{parse_year_month_age()} returns a vector of ages in months.
\item \code{parse_yymm_age()}: returns a vector of ages in months.
}
}
\description{
Convert between age in months, years;months, and yymm age formats
}
\details{
For \code{format_year_month_age()}, ages of \code{NA} return \code{"NA;NA"}.

For \code{parse_year_month_age()}, values that cannot be parsed
return \code{NA}.

This format by default is not numerically ordered. This means that \code{c("2;0", "10;10", "10;9")} would sort as \code{c("10;10", "10;9", "2;0")}. The function
\code{stringr::str_sort(..., numeric = TRUE)} will sort this vector correctly.
}
\examples{
ages <- c(26, 58, 25, 67, 21, 59, 36, 43, 27, 49, NA)
ym_ages <- format_year_month_age(ages)
ym_ages

parse_year_month_age(ym_ages)

parse_yymm_age(c("0204", "0310"))

parse_yymm_age(c("ab_0204", "ab_0310"), start = 4)
}
\concept{data-utils}
