% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss-centiles.R
\name{check_sample_centiles}
\alias{check_sample_centiles}
\title{(Deprecated) Compute the percentage of points under each centile line}
\usage{
check_sample_centiles(
  data,
  model,
  var_x,
  var_y,
  centiles = c(5, 10, 25, 50, 75, 90, 95)
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{check_sample_centiles()} was renamed to \code{check_model_centiles()}.
}
\concept{gamlss}
\keyword{internal}
