% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_features_consonants}
\alias{data_features_consonants}
\alias{data_features_vowels}
\title{Phonetic features of consonants and vowels}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 12 columns.
}
\usage{
data_features_consonants

data_features_vowels
}
\description{
These are two dataframes that contain conventional phonetic features of
the consonants and vowels used by CMU phonetic alphabet.
}
\details{
\subsection{What are phonetic features?}{

We produce most speech sounds by manipulating airflow from the lungs through
the larynx and through the oral or nasal cavities. We need a system for
describing how that airflow is shaped and filtered to make sounds, and phonetic
features are the conventional system for describing sounds. First, let's take
for granted that there is a basic distinction between vowels and consonants. For vowels,
the airflow is almost-always resonant (voicing from the larynx) and largely
unobstructed but tweaked by the position of the tongue.  For
consonants, there is some constriction along the vocal tract.

We describe the consonants by using phonetic features that specify the
constriction in the vocal tract along with the state of larynx:
\itemize{
\item \strong{place of articulation}: where the constriction occurs. If both lips are
used, the place is \emph{bilabial}. When the tongue is involved, the place
describes where the tongue creates the constriction: dental (with the
teeth), alveolar (ridge), velar (velum). And so on.
\item \strong{manner of articulation}: how much constriction occurs and whether anything
else special is going on. Sounds that barely constrict are \emph{approximants}
(or semivowels). Sounds that fully close are \emph{stops}. Sounds
that constrict enough to produce turbulent airflow are \emph{fricatives}.
\emph{Nasal} stops are stops that allow airflow through the nasal cavity.
And so on.
\item \strong{voicing}: whether or not the larynx is closed (and vibrating) to
produce voicing. /s/ and /z/ are two sounds that the same place and
manner of articulation, but differ in their voicing. /s/ is \emph{voiceless},
/z/ is \emph{voiced}.
}

The chart of the \href{https://en.wikipedia.org/wiki/International_Phonetic_Alphabet_chart}{International Phonetic Alphabet}
arranges consonant sounds according to the above three features of place of
articulation (columns), manner of articulation (rows) and voicing (left and
right position in a cell).

Note that all of these features reflect degrees of some tendency. That is,
there is some coherent way in which can order the individual features in a
set of phonetic feature. Place of articulation can be ordered from front
(lips and teeth) to back (tongue root and larynx). Manner can be ordered by
degree of constriction from stops (closure) to fricatives (near closure) to
approximants (much less constructions), but there are secondary manner
features (nasal or lateral airflow, for example) that complicate a
single-dimensional ordering.

I want to point out that phonetic features are descriptive and
classificatory. The features are descriptive: /p/ is \emph{the} voiceless,
bilabial stop. Every Wikipedia article of an individual speech sound(like
\href{https://en.wikipedia.org/wiki/Voiceless_bilabial_plosive}{/p/}) begins by
enumerating its features. On the flipside, the features define natural
classes or classifications/categories of sounds: The "stops" are the
consonants that are made using a complete constriction of airflow at some
point along the vocal tract.

Vowels are described along three

Phonologists (who study how systems of speech sound work) will tack on other
superset features or carve up place and manner into different sets of
distinctions.

(Every statement, I make here is reductive and English centric, so assume there are exception for every tendency I mention. For example, voiceless vowels exist as do sounds that get their airflow without using the lungs.)

Notes on the features chosen:

Most of the features are self-evident and definitional. For example, as I
said above, /p/ is \emph{the} bilabial voiceless stop. For fuzzier features, I consulted the general
IPA chart and the Wikipedia page on English phonology. These issues included
things like: \emph{what are the lax vowels again?} or \emph{the last two rows of the
consonant tables are approximants, so /r,l,j/ are approximants.}

Some features have alternative feature sets in order to accommodate degrees
of aggregation. For example, /r,l,j,w/ are \emph{approximant} in \code{manner}
but divided into \emph{liquid} and \emph{glide} in \code{manner_alt}.
}

\subsection{Consonant features}{

\code{data_features_consonants} is a dataframe with 24 rows and
10 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::kable(data_features_consonants)
}\if{html}{\out{</div>}}\tabular{llllllllll}{
   phone \tab cmubet \tab wiscbet \tab voicing \tab manner \tab manner_alt \tab place \tab place_fct \tab sonorance \tab sonorance_alt \cr
   p \tab P \tab p \tab voiceless \tab stop \tab stop \tab labial \tab labial \tab obstruent \tab obstruent \cr
   b \tab B \tab b \tab voiced \tab stop \tab stop \tab labial \tab labial \tab obstruent \tab obstruent \cr
   t \tab T \tab t \tab voiceless \tab stop \tab stop \tab alveolar \tab alveolar \tab obstruent \tab obstruent \cr
   d \tab D \tab d \tab voiced \tab stop \tab stop \tab alveolar \tab alveolar \tab obstruent \tab obstruent \cr
   k \tab K \tab k \tab voiceless \tab stop \tab stop \tab velar \tab velar \tab obstruent \tab obstruent \cr
   g \tab G \tab g \tab voiced \tab stop \tab stop \tab velar \tab velar \tab obstruent \tab obstruent \cr
   tʃ \tab CH \tab tsh \tab voiceless \tab affricate \tab affricate \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   dʒ \tab JH \tab dzh \tab voiced \tab affricate \tab affricate \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   m \tab M \tab m \tab voiced \tab nasal \tab nasal \tab labial \tab labial \tab sonorant \tab sonorant \cr
   n \tab N \tab n \tab voiced \tab nasal \tab nasal \tab alveolar \tab alveolar \tab sonorant \tab sonorant \cr
   ŋ \tab NG \tab ng \tab voiced \tab nasal \tab nasal \tab velar \tab velar \tab sonorant \tab sonorant \cr
   f \tab F \tab f \tab voiceless \tab fricative \tab fricative \tab labiodental \tab labiodental \tab obstruent \tab strident \cr
   v \tab V \tab v \tab voiced \tab fricative \tab fricative \tab labiodental \tab labiodental \tab obstruent \tab strident \cr
   θ \tab TH \tab th \tab voiceless \tab fricative \tab fricative \tab dental \tab dental \tab obstruent \tab obstruent \cr
   ð \tab DH \tab dh \tab voiced \tab fricative \tab fricative \tab dental \tab dental \tab obstruent \tab obstruent \cr
   s \tab S \tab s \tab voiceless \tab fricative \tab fricative \tab alveolar \tab alveolar \tab obstruent \tab strident \cr
   z \tab Z \tab z \tab voiced \tab fricative \tab fricative \tab alveolar \tab alveolar \tab obstruent \tab strident \cr
   ʃ \tab SH \tab sh \tab voiceless \tab fricative \tab fricative \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   ʒ \tab ZH \tab zh \tab voiced \tab fricative \tab fricative \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   h \tab HH \tab h \tab voiceless \tab fricative \tab fricative \tab glottal \tab glottal \tab obstruent \tab obstruent \cr
   l \tab L \tab l \tab voiced \tab approximant \tab liquid \tab alveolar \tab alveolar \tab sonorant \tab sonorant \cr
   r \tab R \tab r \tab voiced \tab approximant \tab liquid \tab postalveolar \tab postalveolar \tab sonorant \tab sonorant \cr
   w \tab W \tab w \tab voiced \tab approximant \tab glide \tab labiovelar \tab NA \tab sonorant \tab sonorant \cr
   j \tab Y \tab j \tab voiced \tab approximant \tab glide \tab palatal \tab palatal \tab sonorant \tab sonorant \cr
}


Description of each column:

\describe{
\item{phone}{phone in IPA}
\item{cmubet}{phone in the CMU alphabet}
\item{wiscbet}{phone in an older system used by our lab}
\item{voicing}{\emph{voiced} versus \emph{voiceless}}
\item{voicing_alt}{\emph{spread_glottis} versus \emph{plain}}
\item{manner}{manner of articulation}
\item{manner_alt}{alternative manner coding that separates \emph{approximants}
into \emph{liquids} and \emph{glides}}
\item{place}{place of articulation}
\item{place_fct}{place coded as a factor and ordered based on
frontness of the articulators. \emph{labiovelar} is recoded as \code{NA}.}
\item{sonorance}{\emph{obstruent} versus \emph{sonorant}}
\item{sonorance_alt}{\emph{obstruant} versus \emph{sonorant} versus \emph{strident}.}
}

Levels of the factor columns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_features_consonants |>
  lapply(levels) |>
  Filter(length, x = _)
#> $place_fct
#> [1] "labial"       "labiodental"  "dental"       "alveolar"     "postalveolar"
#> [6] "palatal"      "velar"        "glottal"
}\if{html}{\out{</div>}}
}

\subsection{Considerations about consonant features}{

The CMU alphabet does not include a glottal stop.

Here /f,v/ are coded as \emph{strident} following
\href{https://en.wikipedia.org/wiki/Sibilant}{Wikipedia} and
\emph{Sound Pattern of English}. If this feature value doesn't seem right, we
should probably use an alternative feature of \emph{sibilant} for the
stridents minus /f,v/.

The alternative voicing scheme was suggested by a colleague because of how
the voice-voiceless phonetic contrast is achieved with different articulatory
strategies in different languages. Note that \code{voicing_alt} does not assign
a feature to nasals or approximants.
}

\subsection{Vowels}{

\code{data_features_vowels} is a dataframe with 17 rows and
11 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::kable(data_features_vowels)
}\if{html}{\out{</div>}}\tabular{llllllllllll}{
   phone \tab cmubet \tab wiscbet \tab hint \tab manner \tab manner_alt \tab tenseness \tab height \tab height_fct \tab backness \tab backness_fct \tab rounding \cr
   i \tab IY \tab i \tab beat \tab vowel \tab vowel \tab tense \tab high \tab high \tab front \tab front \tab unrounded \cr
   ɪ \tab IH \tab I \tab bit \tab vowel \tab vowel \tab lax \tab high \tab high \tab front \tab front \tab unrounded \cr
   eɪ \tab EY \tab eI \tab bait \tab vowel \tab vowel \tab tense \tab mid \tab mid \tab front \tab front \tab unrounded \cr
   ɛ \tab EH \tab E \tab bet \tab vowel \tab vowel \tab lax \tab mid \tab mid \tab front \tab front \tab unrounded \cr
   æ \tab AE \tab ae \tab bat \tab vowel \tab vowel \tab lax \tab low \tab low \tab front \tab front \tab unrounded \cr
   ʌ \tab AH \tab ^ \tab but \tab vowel \tab vowel \tab lax \tab mid \tab mid \tab central \tab central \tab unrounded \cr
   ə \tab AH \tab 4 \tab comma \tab vowel \tab vowel \tab lax \tab mid \tab mid \tab central \tab central \tab unrounded \cr
   u \tab UW \tab u \tab boot \tab vowel \tab vowel \tab tense \tab high \tab high \tab back \tab back \tab rounded \cr
   ʊ \tab UH \tab U \tab book \tab vowel \tab vowel \tab lax \tab high \tab high \tab back \tab back \tab rounded \cr
   oʊ \tab OW \tab oU \tab boat \tab vowel \tab vowel \tab tense \tab mid \tab mid \tab back \tab back \tab rounded \cr
   ɔ \tab AO \tab c \tab bought \tab vowel \tab vowel \tab tense \tab mid \tab mid \tab back \tab back \tab rounded \cr
   ɑ \tab AA \tab @ \tab bot \tab vowel \tab vowel \tab tense \tab low \tab low \tab back \tab back \tab unrounded \cr
   aʊ \tab AW \tab @U \tab bout \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab NA \tab diphthong \cr
   aɪ \tab AY \tab @I \tab bite \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab NA \tab diphthong \cr
   ɔɪ \tab OY \tab cI \tab boy \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab NA \tab diphthong \cr
   ɝ \tab ER \tab 3^ \tab letter \tab vowel \tab r-colored \tab r-colored \tab mid \tab mid \tab central \tab central \tab r-colored \cr
   ɚ \tab ER \tab 4^ \tab burt \tab vowel \tab r-colored \tab r-colored \tab mid \tab mid \tab central \tab central \tab r-colored \cr
}


Description of each column:

\describe{
\item{phone}{phone in IPA}
\item{cmubet}{phone in the CMU alphabet}
\item{wiscbet}{phone in an older system used by our lab}
\item{hint}{a word containing the selected vowel}
\item{manner}{manner of articulation}
\item{manner_alt}{alternative manner with \emph{vowel}, \emph{diphthong} and \emph{r-colored}}
\item{tenseness}{\emph{tense} versus \emph{lax} (versus \emph{diphthong} and \emph{r-colored})}
\item{height}{vowel height}
\item{height_fct}{height coded as a factor ordered \emph{high},
\emph{mid}, \emph{low}. \emph{diphthong} is recoded to \code{NA}.}
\item{backness}{vowel backness}
\item{backness_fct}{backness coded as a factor ordered \emph{front},
\emph{central}, \emph{back}. \emph{diphthong} is recoded to \code{NA}.}
\item{rounding}{\emph{unrounded} versus \emph{rounded} (versus \emph{diphthong} and
\emph{r-colored})}
}

Levels of the factor columns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_features_vowels |>
  lapply(levels) |>
  Filter(length, x = _)
#> $height_fct
#> [1] "high" "mid"  "low" 
#> 
#> $backness_fct
#> [1] "front"   "central" "back"
}\if{html}{\out{</div>}}
}

\subsection{Considerations about vowel features}{

I don't consider /eɪ/ and /oʊ/ to be diphthongs, but perhaps \code{manner_alt}
could encode the difference of these vowels from the others.

In the CMU alphabet and ARPAbet, vowels can include a number to indicate
vowel stress, so \code{AH1} or \code{AH2} is /ʌ/ but \code{AH0} is /ə/.

The vowel features for General American English,
\href{https://en.wikipedia.org/wiki/English_phonology}{according to Wikipedia},
are as follows:\if{html}{\out{
<table style="text-align: center;">
<tbody>
<tr class="header">
<th rowspan="2"></th>
<th colspan="2">Front</th>
<th colspan="2">Central</th>
<th colspan="2">Back</th>
</tr>
<tr class="odd">
<th>lax</th>
<th>tense</th>
<th>lax</th>
<th>tense</th>
<th>lax</th>
<th>tense</th>
</tr>
}}
\if{html}{\out{
<tr class="odd">
<th>Close</th>
<td>ɪ</td>
<td>i</td>
<td></td>
<td></td>
<td>ʊ</td>
<td>u</td>
</tr>
<tr class="even">
<th>Mid</th>
<td>ɛ</td>
<td>eɪ</td>
<td>ə</td>
<td>(ɜ)</td>
<td></td>
<td>oʊ</td>
</tr>
<tr class="odd">
<th>Open</th>
<td>æ</td>
<td></td>
<td>(ʌ)</td>
<td>ɑ</td>
<td></td>
<td>(ɔ)</td>
</tr>
<tr class="even">
<th>Diphthongs</th>
<td colspan="6">aɪ ɔɪ aʊ</td>
</tr>
</tbody>
</table>
}}


I adapted these features in this way:
\itemize{
\item \emph{tense} and \emph{lax} features were directly borrowed. Diphthongs and
r-colored vowels are were not assign a tenseness.
\item /ʌ,ɔ/ raised to \emph{mid} (following the general IPA chart)
\item /ɑ/ moved to \emph{back} (following the general IPA)
\item diphthongs have no backness or height
\item r-colored vowels were given the backness and height of the /ʌ,ə/
}

Based on the assumption that /ʌ,ə/ are the same general vowel with
differing stress, these vowels have the same features. This definition
clashes with the general IPA chart which places /ʌ/ as a back vowel. However,
/ʌ/ is a conventional notation. Quoting
\href{https://en.wikipedia.org/wiki/English_phonology}{Wikipedia} again:
"Although the notation /ʌ/ is used for the vowel of STRUT in RP and General
American, the actual pronunciation is closer to a near-open central
vowel [ɐ] in RP and advanced back [ʌ̟] in General American." That is,
/ʌ/ is fronted in American English (hence, \emph{mid}) in American English.
}
}
\concept{datasets}
\keyword{datasets}
