% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_features_consonants}
\alias{data_features_consonants}
\alias{data_features_vowels}
\title{Phonetic features of consonants and vowels}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 11 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 13 columns.
}
\usage{
data_features_consonants

data_features_vowels
}
\description{
This package provides dataframes of information about the consonants and
vowels in American English. The \emph{phonetic} features are conventional
descriptions of how sounds are produced. See also
\link{data_acq_consonants}.
}
\details{
Most of the phonetic features are self-evident and definitional. For example, /p/ is
\emph{the} bilabial voiceless stop. For fuzzier features, I consulted the general
IPA chart and the Wikipedia page on English phonology. These issues included
things like: \emph{what are the lax vowels again?} or \emph{the last two rows of the
consonant tables are approximants, so /r,l,j/ are approximants.}

Some features have alternative feature sets in order to accommodate degrees
of aggregation. For example, /r,l,j,w/ are \emph{approximant} in \code{manner}
but divided into \emph{liquid} and \emph{glide} in \code{manner_alt}.
\subsection{Phonetic features of consonants}{

\code{data_features_consonants} is a dataframe with 24 rows and
10 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::kable(data_features_consonants)
}\if{html}{\out{</div>}}\tabular{lllllllllll}{
   phone \tab cmubet \tab wiscbet \tab voicing \tab voicing_alt \tab manner \tab manner_alt \tab place \tab place_fct \tab sonorance \tab sonorance_alt \cr
   p \tab P \tab p \tab voiceless \tab spread_glottis \tab stop \tab stop \tab labial \tab labial \tab obstruent \tab obstruent \cr
   b \tab B \tab b \tab voiced \tab plain \tab stop \tab stop \tab labial \tab labial \tab obstruent \tab obstruent \cr
   t \tab T \tab t \tab voiceless \tab spread_glottis \tab stop \tab stop \tab alveolar \tab alveolar \tab obstruent \tab obstruent \cr
   d \tab D \tab d \tab voiced \tab plain \tab stop \tab stop \tab alveolar \tab alveolar \tab obstruent \tab obstruent \cr
   k \tab K \tab k \tab voiceless \tab spread_glottis \tab stop \tab stop \tab velar \tab velar \tab obstruent \tab obstruent \cr
   g \tab G \tab g \tab voiced \tab plain \tab stop \tab stop \tab velar \tab velar \tab obstruent \tab obstruent \cr
   tʃ \tab CH \tab tsh \tab voiceless \tab spread_glottis \tab affricate \tab affricate \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   dʒ \tab JH \tab dzh \tab voiced \tab plain \tab affricate \tab affricate \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   m \tab M \tab m \tab voiced \tab NA \tab nasal \tab nasal \tab labial \tab labial \tab sonorant \tab sonorant \cr
   n \tab N \tab n \tab voiced \tab NA \tab nasal \tab nasal \tab alveolar \tab alveolar \tab sonorant \tab sonorant \cr
   ŋ \tab NG \tab ng \tab voiced \tab NA \tab nasal \tab nasal \tab velar \tab velar \tab sonorant \tab sonorant \cr
   f \tab F \tab f \tab voiceless \tab spread_glottis \tab fricative \tab fricative \tab labiodental \tab labiodental \tab obstruent \tab strident \cr
   v \tab V \tab v \tab voiced \tab plain \tab fricative \tab fricative \tab labiodental \tab labiodental \tab obstruent \tab strident \cr
   θ \tab TH \tab th \tab voiceless \tab spread_glottis \tab fricative \tab fricative \tab dental \tab dental \tab obstruent \tab obstruent \cr
   ð \tab DH \tab dh \tab voiced \tab plain \tab fricative \tab fricative \tab dental \tab dental \tab obstruent \tab obstruent \cr
   s \tab S \tab s \tab voiceless \tab spread_glottis \tab fricative \tab fricative \tab alveolar \tab alveolar \tab obstruent \tab strident \cr
   z \tab Z \tab z \tab voiced \tab plain \tab fricative \tab fricative \tab alveolar \tab alveolar \tab obstruent \tab strident \cr
   ʃ \tab SH \tab sh \tab voiceless \tab spread_glottis \tab fricative \tab fricative \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   ʒ \tab ZH \tab zh \tab voiced \tab plain \tab fricative \tab fricative \tab postalveolar \tab postalveolar \tab obstruent \tab strident \cr
   h \tab HH \tab h \tab voiceless \tab spread_glottis \tab fricative \tab fricative \tab glottal \tab glottal \tab obstruent \tab obstruent \cr
   l \tab L \tab l \tab voiced \tab NA \tab approximant \tab liquid \tab alveolar \tab alveolar \tab sonorant \tab sonorant \cr
   r \tab R \tab r \tab voiced \tab NA \tab approximant \tab liquid \tab postalveolar \tab postalveolar \tab sonorant \tab sonorant \cr
   w \tab W \tab w \tab voiced \tab NA \tab approximant \tab glide \tab labiovelar \tab NA \tab sonorant \tab sonorant \cr
   j \tab Y \tab j \tab voiced \tab NA \tab approximant \tab glide \tab palatal \tab palatal \tab sonorant \tab sonorant \cr
}


Description of each column:

\describe{
\item{phone}{phone in IPA}
\item{cmubet}{phone in the CMU alphabet}
\item{wiscbet}{phone in an older system used by our lab}
\item{voicing}{\emph{voiced} versus \emph{voiceless}}
\item{voicing_alt}{\emph{spread_glottis} versus \emph{plain}}
\item{manner}{manner of articulation}
\item{manner_alt}{alternative manner coding that separates \emph{approximants}
into \emph{liquids} and \emph{glides}}
\item{place}{place of articulation}
\item{place_fct}{place coded as a factor and ordered based on
frontness of the articulators. \emph{labiovelar} is recoded as \code{NA}.}
\item{sonorance}{\emph{obstruent} versus \emph{sonorant}}
\item{sonorance_alt}{\emph{obstruant} versus \emph{sonorant} versus \emph{strident}.}
}

Levels of the factor columns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_features_consonants |>
  lapply(levels) |>
  Filter(length, x = _)
#> $place_fct
#> [1] "labial"       "labiodental"  "dental"       "alveolar"     "postalveolar"
#> [6] "palatal"      "velar"        "glottal"
}\if{html}{\out{</div>}}
}

\subsection{Considerations about consonant phonetic features}{

The CMU alphabet does not include a glottal stop.

Here /f,v/ are coded as \emph{strident} following
\href{https://en.wikipedia.org/wiki/Sibilant}{Wikipedia} and
\emph{Sound Pattern of English}. If this feature value doesn't seem right, we
should probably use an alternative feature of \emph{sibilant} for the
stridents minus /f,v/.

The alternative voicing scheme was suggested by a colleague because of how
the voice-voiceless phonetic contrast is achieved with different articulatory
strategies in different languages. Note that \code{voicing_alt} does not assign
a feature to nasals or approximants.
}

\subsection{Phonetic features of vowels}{

\code{data_features_vowels} is a dataframe with 17 rows and
11 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::kable(data_features_vowels)
}\if{html}{\out{</div>}}\tabular{lllllllllllll}{
   phone \tab cmubet \tab wiscbet \tab hint \tab manner \tab manner_alt \tab tenseness \tab height \tab height_fct \tab height_alt \tab backness \tab backness_fct \tab rounding \cr
   i \tab IY \tab i \tab beat \tab vowel \tab vowel \tab tense \tab high \tab high \tab high \tab front \tab front \tab unrounded \cr
   ɪ \tab IH \tab I \tab bit \tab vowel \tab vowel \tab lax \tab mid-high \tab mid-high \tab high \tab front \tab front \tab unrounded \cr
   eɪ \tab EY \tab eI \tab bait \tab vowel \tab vowel \tab tense \tab mid-high \tab mid-high \tab mid \tab front \tab front \tab unrounded \cr
   ɛ \tab EH \tab E \tab bet \tab vowel \tab vowel \tab lax \tab mid-low \tab mid-low \tab mid \tab front \tab front \tab unrounded \cr
   æ \tab AE \tab ae \tab bat \tab vowel \tab vowel \tab lax \tab low \tab low \tab low \tab front \tab front \tab unrounded \cr
   ʌ \tab AH \tab ^ \tab but \tab vowel \tab vowel \tab lax \tab mid-low \tab mid-low \tab mid \tab central \tab central \tab unrounded \cr
   ə \tab AH \tab 4 \tab comma \tab vowel \tab vowel \tab lax \tab mid-low \tab mid-low \tab mid \tab central \tab central \tab unrounded \cr
   u \tab UW \tab u \tab boot \tab vowel \tab vowel \tab tense \tab high \tab high \tab high \tab back \tab back \tab rounded \cr
   ʊ \tab UH \tab U \tab book \tab vowel \tab vowel \tab lax \tab mid-high \tab mid-high \tab high \tab back \tab back \tab rounded \cr
   oʊ \tab OW \tab oU \tab boat \tab vowel \tab vowel \tab tense \tab mid-high \tab mid-high \tab mid \tab back \tab back \tab rounded \cr
   ɔ \tab AO \tab c \tab bought \tab vowel \tab vowel \tab tense \tab mid-low \tab mid-low \tab low \tab back \tab back \tab rounded \cr
   ɑ \tab AA \tab @ \tab bot \tab vowel \tab vowel \tab tense \tab low \tab low \tab low \tab back \tab back \tab unrounded \cr
   aʊ \tab AW \tab @U \tab bout \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab diphthong \tab NA \tab diphthong \cr
   aɪ \tab AY \tab @I \tab bite \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab diphthong \tab NA \tab diphthong \cr
   ɔɪ \tab OY \tab cI \tab boy \tab vowel \tab diphthong \tab diphthong \tab diphthong \tab NA \tab diphthong \tab diphthong \tab NA \tab diphthong \cr
   ɝ \tab ER \tab 3^ \tab letter \tab vowel \tab r-colored \tab r-colored \tab mid-low \tab mid-low \tab mid \tab central \tab central \tab r-colored \cr
   ɚ \tab ER \tab 4^ \tab burt \tab vowel \tab r-colored \tab r-colored \tab mid-low \tab mid-low \tab mid \tab central \tab central \tab r-colored \cr
}


Description of each column:

\describe{
\item{phone}{phone in IPA}
\item{cmubet}{phone in the CMU alphabet}
\item{wiscbet}{phone in an older system used by our lab}
\item{hint}{a word containing the selected vowel}
\item{manner}{manner of articulation}
\item{manner_alt}{alternative manner with \emph{vowel}, \emph{diphthong} and \emph{r-colored}}
\item{tenseness}{\emph{tense} versus \emph{lax} (versus \emph{diphthong} and \emph{r-colored})}
\item{height}{vowel height on a four-level scale}
\item{height_fct}{height coded as a factor ordered \emph{high},
\emph{mid-high}, \emph{mid-low}, \emph{low}. \emph{diphthong} is recoded to \code{NA}.}
\item{height_alt}{vowel height on a three-level scale}
\item{backness}{vowel backness}
\item{backness_fct}{backness coded as a factor ordered \emph{front},
\emph{central}, \emph{back}. \emph{diphthong} is recoded to \code{NA}.}
\item{rounding}{\emph{unrounded} versus \emph{rounded} (versus \emph{diphthong} and
\emph{r-colored})}
}

Levels of the factor columns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_features_vowels |>
  lapply(levels) |>
  Filter(length, x = _)
#> $height_fct
#> [1] "high"     "mid-high" "mid-low"  "low"     
#> 
#> $backness_fct
#> [1] "front"   "central" "back"
}\if{html}{\out{</div>}}
}

\subsection{Considerations about vowel features}{

I don't consider /eɪ/ and /oʊ/ to be diphthongs, but perhaps \code{manner_alt}
could encode the difference of these vowels from the others.

In the CMU alphabet and ARPAbet, vowels can include a number to indicate
vowel stress, so \code{AH1} or \code{AH2} is /ʌ/ but \code{AH0} is /ə/.

The vowel features for General American English,
\href{https://en.wikipedia.org/wiki/English_phonology}{according to Wikipedia},
are as follows:\if{html}{\out{
<table style="text-align: center;">
<tbody>
<tr class="header">
<th rowspan="2"></th>
<th colspan="2">Front</th>
<th colspan="2">Central</th>
<th colspan="2">Back</th>
</tr>
<tr class="odd">
<th>lax</th>
<th>tense</th>
<th>lax</th>
<th>tense</th>
<th>lax</th>
<th>tense</th>
</tr>
}}
\if{html}{\out{
<tr class="odd">
<th>Close</th>
<td>ɪ</td>
<td>i</td>
<td></td>
<td></td>
<td>ʊ</td>
<td>u</td>
</tr>
<tr class="even">
<th>Mid</th>
<td>ɛ</td>
<td>eɪ</td>
<td>ə</td>
<td>(ɜ)</td>
<td></td>
<td>oʊ</td>
</tr>
<tr class="odd">
<th>Open</th>
<td>æ</td>
<td></td>
<td>(ʌ)</td>
<td>ɑ</td>
<td></td>
<td>(ɔ)</td>
</tr>
<tr class="even">
<th>Diphthongs</th>
<td colspan="6">aɪ ɔɪ aʊ</td>
</tr>
</tbody>
</table>
}}


I adapted these features in this way:
\itemize{
\item A four-level breakdown of height (high, mid-high, mid-low, low) was used
instead of a three-level one (close, mid, open).
\item \emph{tense} and \emph{lax} features were directly borrowed. Diphthongs and
r-colored vowels are were not assign a tenseness.
\item /ɑ/ moved to \emph{back} (following the general IPA)
\item diphthongs have no backness or height
\item r-colored vowels were given the backness and height of the /ʌ,ə/
}

Based on the assumption that /ʌ,ə/ are the same general vowel with
differing stress, these vowels have the same features. This definition
clashes with the general IPA chart which places /ʌ/ as a back vowel. However,
/ʌ/ is a conventional notation. Quoting
\href{https://en.wikipedia.org/wiki/English_phonology}{Wikipedia} again:
"Although the notation /ʌ/ is used for the vowel of STRUT in RP and General
American, the actual pronunciation is closer to a near-open central
vowel [ɐ] in RP and advanced back [ʌ̟] in General American." That is,
/ʌ/ is fronted in American English (hence, \emph{mid}) in American English.
}
}
\concept{datasets}
\keyword{datasets}
