% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tocs_items}
\alias{data_tocs_items}
\title{Items from the TOCS+ used by the WISC Lab}
\format{
A data frame with 103 rows and 4 variables:
\describe{
\item{tocs_type}{whether the item is a \code{single-word} or \code{multiword} utterance}
\item{tocs_level}{number of words in the prompt}
\item{tocs_item}{identifier for the item (e.g., within filenames)}
\item{tocs_prompt}{the actual prompt presented to the child}
\item{tocs_item_note}{any notes about an item's usage in experiments or analyses}
}
}
\usage{
data_tocs_items
}
\description{
We use items from the Test of Children's Speech (TOCS+,
Hodge & Gotzke, 2014) in our child speech and listener intelligibility
studies.
}
\details{
Most of these items are subset from a larger pool of items on the TOCS+.
Hustad and colleagues (2021) is a supplemental material for a larger study
about speech intelligibility which describes where the items were taken from
within the larger TOCS+ task. (It is unclear, as I write this in 2026, which
items were added by our group to the TOCS+ pool, so I have added "possibly
not an original TOCS item" to items that do not appear in that supplemental
material.)

In this picture-prompted repetition task, a child hears a \emph{prompt},
repeats it (\emph{production}), and later on listeners transcribe it
(\emph{transcription}). In our listener-response database, we differentiate these
three forms with fields named \code{tocs_prompt} (prompt), \code{sentence}
(production), and \code{response} (transcription).
}
\references{
Hodge, M. M., & Gotzke, C. (2014). Construct-related validity of
the TOCS+ measures: Comparison of intelligibility and speaking rate scores
in children with and without speech disorders. \emph{Journal of Communication
Disorders}, \emph{51}, 51–63. \url{https://doi.org/10.1016/j.jcomdis.2014.06.007}

Hustad, K. C., Mahr, T. J., Natzke, P., & Rathouz, P. J. (2021).
Supplemental Material 1: Intelligibility growth between 30 and 119 months
(Hustad et al., 2021) (Version 1). ASHA Journals.
\url{https://doi.org/10.23641/asha.16583426.v1}
}
\concept{datasets}
\keyword{datasets}
