% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{join_to_split}
\alias{join_to_split}
\title{Join data onto resampled IDs}
\usage{
join_to_split(x, y, by, validate = FALSE)
}
\arguments{
\item{x}{an rset object created by \code{rsample::bootstraps()}}

\item{y}{y dataframe with a column of the id values which was resampled to
create \code{x}}

\item{by}{the name of column in y with the data}

\item{validate}{whether to validate the join by counting the number of rows
associated with each id. Defaults to \code{FALSE}.}
}
\value{
the original rset object with its \code{x$data} updated to join with \code{y}
and with the row numbers \code{x$in_id} updated to work on the expanded dataset.
}
\description{
Join data onto resampled IDs
}
\examples{
library(dplyr)
data_trees <- tibble::as_tibble(datasets::Orange)

data_tree_ids <- distinct(data_trees, Tree)

# Resample ids
data_bootstraps <- data_tree_ids \%>\%
  rsample::bootstraps(times = 20) \%>\%
  rename(splits_id = splits) \%>\%
  # Attach data to resampled ids
  mutate(
    data_splits = splits_id \%>\% purrr::map(
      join_to_split,
      data_trees,
      by = "Tree",
      validate = TRUE
    )
  )

data_bootstraps
}
\concept{etc}
